Move different lock managers into separate files
